Rename rectify.py to activations.py
