+++ fastats/maths/activations.py
